begin architecture rewrite
