+++ launch.py
@@ -4,4 +4,4 @@ dotenv.load_dotenv()
 
 from src.bot import bot
 
-bot.start()
+bot.start(dotenv.get_key(".env", "TOKEN"))
